# docs/api-spec.json
{
  "components": {
    "schemas": {
      "AddBody": {
        "properties": {
          "stock_id": {
            "title": "Stock Id",
            "type": "string"
          }
        },
        "required": [
          "stock_id"
        ],
        "title": "AddBody",
        "type": "object"
      },
...
      "BulkAddBody": {
        "properties": {
          "stock_ids": {
            "items": {
              "type": "string"
            },
            "title": "Stock Ids",
            "type": "array"
          }
        },
        "required": [
          "stock_ids"
        ],
        "title": "BulkAddBody",
        "type": "object"
      },
      "BulkRemoveBody": {
        "properties": {
          "stock_ids": {
            "items": {
              "type": "string"
            },
            "title": "Stock Ids",
            "type": "array"
          }
        },
        "required": [
          "stock_ids"
        ],
        "title": "BulkRemoveBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TagsUpdateBody": {
        "description": "覆寫單檔的所有 tags。空 list = 清空。\n後端會做 trim + 去重（保序），UI 不必預先處理。",
        "properties": {
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          }
        },
        "required": [
          "tags"
        ],
        "title": "TagsUpdateBody",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WatchlistEntry": {
        "description": "回應用 — CamelModel 會自動 serialize 成 camelCase (stockId / stockName)。",
        "properties": {
          "stockId": {
            "title": "Stockid",
            "type": "string"
          },
          "stockName": {
            "title": "Stockname",
            "type": "string"
          },
          "tags": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          }
        },
        "required": [
          "stockId",
          "stockName"
        ],
        "title": "WatchlistEntry",
        "type": "object"
      },
      "WatchlistLookup": {
        "properties": {
          "stockId": {
            "title": "Stockid",
            "type": "string"
          },
          "stockName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stockname"
          }
        },
        "required": [
          "stockId"
        ],
        "title": "WatchlistLookup",
        "type": "object"
      },
      "WatchlistMover": {
        "properties": {
          "changePct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Changepct"
          },
          "close": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Close"
          },
          "compositeScore": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compositescore"
          },
          "market": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market"
          },
          "stockId": {
            "title": "Stockid",
            "type": "string"
          },
          "stockName": {
            "title": "Stockname",
            "type": "string"
          }
        },
        "required": [
          "stockId",
          "stockName"
        ],
        "title": "WatchlistMover",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Next.js 前端的 FastAPI 後端，包裝既有 app.* 模組。",
    "title": "台股研究儀表板 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/watchlist": {
      "get": {
        "operationId": "list_all_api_watchlist_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WatchlistEntry"
                  },
                  "title": "Response List All Api Watchlist Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List All",
        "tags": [
          "watchlist"
        ]
      },
      "post": {
        "description": "新增自選：名稱一律以 stock_info 為準。\n驗證：股票必須在 stock_info 或 daily_price 表中至少存在一筆資料才能加入，\n否則會寫入垃圾代號（如 9999）導致雷達 / 評分異常。",
        "operationId": "add_api_watchlist_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Add Api Watchlist Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add",
        "tags": [
          "watchlist"
        ]
      }
    },
    "/api/watchlist/bulk-add": {
      "post": {
        "description": "批次新增：若輸入只含代號，會自動從 stock_info 帶出名稱。",
        "operationId": "bulk_add_api_watchlist_bulk_add_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkAddBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Bulk Add Api Watchlist Bulk Add Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bulk Add",
        "tags": [
          "watchlist"
        ]
      }
    },
    "/api/watchlist/bulk-remove": {
      "post": {
        "operationId": "bulk_remove_api_watchlist_bulk_remove_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkRemoveBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Bulk Remove Api Watchlist Bulk Remove Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bulk Remove",
        "tags": [
          "watchlist"
        ]
      }
    },
    "/api/watchlist/lookup/{stock_id}": {
      "get": {
        "description": "代號 → 名稱（供新增表單 auto-fill）。回傳 camelCase (stockName)。",
        "operationId": "lookup_api_watchlist_lookup__stock_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "stock_id",
            "required": true,
            "schema": {
              "title": "Stock Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistLookup"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Lookup",
        "tags": [
          "watchlist"
        ]
      }
    },
    "/api/watchlist/movers": {
      "get": {
        "description": "自選股當日漲/跌幅排行。direction=up|down。",
        "operationId": "movers_api_watchlist_movers_get",
        "parameters": [
          {
            "in": "query",
            "name": "top",
            "required": false,
            "schema": {
              "default": 5,
              "title": "Top",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "direction",
            "required": false,
            "schema": {
              "default": "up",
              "title": "Direction",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WatchlistMover"
                  },
                  "title": "Response Movers Api Watchlist Movers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Movers",
        "tags": [
          "watchlist"
        ]
      }
    },
...
    "/api/watchlist/{stock_id}": {
      "delete": {
        "description": "冪等刪除：刪不存在的代號也回 204（與 portfolio 的 DELETE 一致，retry 安全）。",
        "operationId": "remove_api_watchlist__stock_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "stock_id",
            "required": true,
            "schema": {
              "title": "Stock Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove",
        "tags": [
          "watchlist"
        ]
      }
    },
    "/api/watchlist/{stock_id}/tags": {
      "put": {
        "description": "覆寫單檔 tags。stock_id 不在自選清單 → 404。",
        "operationId": "update_tags_api_watchlist__stock_id__tags_put",
        "parameters": [
          {
            "in": "path",
            "name": "stock_id",
            "required": true,
            "schema": {
              "title": "Stock Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagsUpdateBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistEntry"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Tags",
        "tags": [
          "watchlist"
        ]
      }
    },
...
  }
}